Create a reusable SignInViaGithubModal in the app
